BREAKING CHANGE: refactor Handler to achieve polymorphism
In the past we used struct literal like `&ClosureHandler{}` as the constructor. Now use `NewClosureHandler()` instead. Every parameter's meaning and order matches that previously in struct literal.

Every usage are always interface Handler orientated, struct ClosureHandler is an open helper that implements the interface, therefore it has public member only. But that's
not mandatory, as new common parts are pulled as ClosureMatcherAndParser, I prefer introducing a stable New, comparing to changing every occurrence of the struct literal constructor.